created QuestionButton for questions and refactored logout and settings buttons
