+++ src/components/Hello.jsx
@@ -31,7 +31,7 @@ function AudioPlayer() {
     const data = await createVoice(text);
     return data;
   };
-  console.log("hello")
+  console.log("hello");
 
   return (
     <div>
