--- src/components/button.css
@@ -1,18 +1,0 @@
-.button-wrapper {
-  background-color: #ee9f90;
-  padding: 10px;
-  border-radius: 15px;
-  cursor: pointer;
-  border: none;
-  color: white;
-  font-weight: bold;
-  font-size: 20px;
-  transition-property: all;
-  transition-duration: 200ms;
-  transition-timing-function: ease;
-}
-
-.button-wrapper:hover {
-  background-color: #ffe7e6;
-  color: #ee9f90;
-}
