+++ src/components/buttons.css
@@ -1,0 +1,58 @@
+.button-wrapper {
+  background-color: #ee9f90;
+  padding: 10px;
+  border-radius: 15px;
+  cursor: pointer;
+  border: none;
+  color: white;
+  font-weight: bold;
+  font-size: 20px;
+  transition-property: all;
+  transition-duration: 200ms;
+  transition-timing-function: ease;
+}
+
+.button-wrapper:hover {
+  background-color: #ffe7e6;
+  color: #ee9f90;
+}
+
+.question-button-wrapper {
+  background-color: #ee9f90;
+  padding: 10px;
+  border-radius: 15px;
+  cursor: pointer;
+  border-style: solid;
+  border-width: 3px 3px 8px;
+  border-color: #9f7b71;
+  color: white;
+  font-weight: bold;
+  font-size: 20px;
+  height: 150px;
+  width: 150px;
+}
+
+.settings-btn {
+  background-color: #ee9f90;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 40px;
+  height: 40px;
+  padding: 10px;
+  border-radius: 15px;
+}
+
+.log-button {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 40px;
+  height: 40px;
+  padding: 10px;
+  background-color: #ee9f90;
+  border-radius: 15px;
+  cursor: pointer;
+}
