+++ src/components/lessonTest.css
@@ -1,0 +1,12 @@
+.lesson-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.lesson-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 60%;
+}
