--- src/components/logButton.css
@@ -1,12 +1,0 @@
-.log-button {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 40px;
-  height: 40px;
-  padding: 10px;
-  background-color: #ee9f90;
-  border-radius: 15px;
-  cursor: pointer;
-}
