--- src/components/settingsButton.css
@@ -1,11 +1,0 @@
-.settings-btn {
-  background-color: #ee9f90;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 40px;
-  height: 40px;
-  padding: 10px;
-  border-radius: 15px;
-}
